% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_hierachy.R
\name{compact}
\alias{compact}
\title{Compact a set of hexagons of the same resolution into a set of hexagons across multiple levels.}
\usage{
compact(h3_indexes)
}
\arguments{
\item{h3_indexes}{character vector of H3 indexes}
}
\description{
Compact a set of hexagons of the same resolution into a set of hexagons across multiple levels.
}
