% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge-boundary-to-sf.R
\name{edge_boundary_to_sf}
\alias{edge_boundary_to_sf}
\title{Parse edge coordiantes to \code{sf} object}
\usage{
edge_boundary_to_sf(latlng)
}
\arguments{
\item{latlng}{matrix; edge coordinates}
}
\description{
Parse edge coordiantes to \code{sf} object
}
