% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{edge_length}
\alias{edge_length}
\title{Average hexagon edge length in meters or kilometers at the given resolution.}
\usage{
edge_length(res, unit)
}
\arguments{
\item{res}{resolution between 0 and 15}

\item{unit}{either \code{m} or \code{km}}
}
\value{
numeric vector
}
\description{
Average hexagon edge length in meters or kilometers at the given resolution.
}
