% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-boundary-to-sf.R
\name{geo_boundary_to_sf}
\alias{geo_boundary_to_sf}
\alias{geo_boundary_to_sf.matrix}
\alias{geo_boundary_to_sf.list}
\title{Parse geo boundaries to object of class \code{sf}}
\usage{
geo_boundary_to_sf(latlng)

\method{geo_boundary_to_sf}{matrix}(latlng)

\method{geo_boundary_to_sf}{list}(latlng)
}
\arguments{
\item{latlng}{geo coordinates}
}
\description{
Parse geo boundaries to object of class \code{sf}
}
