% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_indexing.R
\name{geo_to_h3}
\alias{geo_to_h3}
\alias{geo_to_h3.numeric}
\alias{geo_to_h3.matrix}
\alias{geo_to_h3.data.frame}
\alias{geo_to_h3.sf}
\title{Index geo-coordinates at the given resolution into H3 addresses.}
\usage{
geo_to_h3(latlng, res)

\method{geo_to_h3}{numeric}(latlng, res = 7)

\method{geo_to_h3}{matrix}(latlng, res = 7)

\method{geo_to_h3}{data.frame}(latlng, res = 7)

\method{geo_to_h3}{sf}(latlng, res = 7)
}
\arguments{
\item{latlng}{geo-coordinates as lat/lng pairs or object of class \code{sf}}

\item{res}{resolution between 0 and 15}
}
\value{
character vector
}
\description{
Index geo-coordinates at the given resolution into H3 addresses.
}
\examples{
road_safety_greater_manchester[1:3, ] \%>\%
  geo_to_h3()
}
