% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-to-sf.R
\name{geo_to_sf}
\alias{geo_to_sf}
\alias{geo_to_sf.matrix}
\title{#' Parse centers/geo-coordinates to object of class \code{sf}.}
\usage{
geo_to_sf(latlng)

\method{geo_to_sf}{matrix}(latlng)
}
\arguments{
\item{latlng}{geo-coordinates}
}
\description{
#' Parse centers/geo-coordinates to object of class \code{sf}.
}
