% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_h3_unidirectional_edge}
\alias{get_h3_unidirectional_edge}
\title{Get the H3 edge index based on the given origin and destination hexagons.}
\usage{
get_h3_unidirectional_edge(origin, destination)
}
\arguments{
\item{origin}{character scalar; origin H3 index}

\item{destination}{character scalar; destination H3 index}
}
\description{
Get the H3 edge index based on the given origin and destination hexagons.
}
