% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_uniedge.R
\name{get_h3_unidirectional_edge_boundary}
\alias{get_h3_unidirectional_edge_boundary}
\title{Get the coordinates defining the unidirectional edge.}
\usage{
get_h3_unidirectional_edge_boundary(h3_edge_index)
}
\arguments{
\item{h3_edge_index}{character scalar; H3 edge index}
}
\description{
Get the coordinates defining the unidirectional edge.
}
