% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge-boundary-to-sf.R
\name{get_h3_unidirectional_edge_boundary_sf}
\alias{get_h3_unidirectional_edge_boundary_sf}
\title{Get the unidirectional edge as \code{sf} object.}
\usage{
get_h3_unidirectional_edge_boundary_sf(h3_edge_index)
}
\description{
Get the unidirectional edge as \code{sf} object.
}
