% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_uniedge.R
\name{get_h3_unidirectional_edges_from_hexagon}
\alias{get_h3_unidirectional_edges_from_hexagon}
\title{Get all of the unidirectional edges from the given H3 index.}
\usage{
get_h3_unidirectional_edges_from_hexagon(h3_index)
}
\arguments{
\item{h3_index}{character scalar; H3 index}
}
\description{
Get all of the unidirectional edges from the given H3 index.
}
\examples{
edges <- road_safety_greater_manchester[1, ] \%>\%
  geo_to_h3() \%>\%
  get_h3_unidirectional_edges_from_hexagon()

h3_unidirectional_edge_is_valid(edges)
}
