% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_uniedge.R
\name{get_origin_h3_index_from_unidirectional_edge}
\alias{get_origin_h3_index_from_unidirectional_edge}
\title{Get the origin hexagon from the unidirectional H3 edge index.}
\usage{
get_origin_h3_index_from_unidirectional_edge(h3_edge_index)
}
\arguments{
\item{h3_edge_index}{character scalar; H3 edge index}
}
\description{
Get the origin hexagon from the unidirectional H3 edge index.
}
