% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{h3_cell_azimuth}
\alias{h3_cell_azimuth}
\title{Returns geodesic azimuth for a cell by its hex index}
\usage{
h3_cell_azimuth(h3_index)
}
\arguments{
\item{h3_index}{string index of the cell}
}
\value{
double value of the azimuth
}
\description{
Returns geodesic azimuth for a cell by its hex index
}
