% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{h3_cell_vecinity}
\alias{h3_cell_vecinity}
\title{Returns cell neighbours of defined order and less with the cell itself}
\usage{
h3_cell_vecinity(h3s, radius)
}
\arguments{
\item{radius}{}
}
\value{
cell neighbors of defined order and less with the cell itself
}
\description{
Returns cell neighbours of defined order and less with the cell itself
}
