% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{h3_distance}
\alias{h3_distance}
\title{Get the grid distance between H3 addresses.}
\usage{
h3_distance(origin, destinations)
}
\arguments{
\item{origin}{character; origin H3 index}

\item{destination}{character vector of H3 destination indexes}
}
\value{
numeric vector
}
\description{
Get the grid distance between H3 addresses.
}
