% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{h3_flow_acc}
\alias{h3_flow_acc}
\title{Calculates water flow accumulation from flow direction table}
\usage{
h3_flow_acc(ifrom, ito)
}
\arguments{
\item{ifrom}{vector of H3 indexes}

\item{ito}{corresponding heights}
}
\value{
map of H3 indexes and cell's accumulation value
}
\description{
Calculates water flow accumulation from flow direction table
}
