% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{h3_flow_dir}
\alias{h3_flow_dir}
\title{Calculates water flow directions and fills depressions
needs h3 indexes and corresponding heights as well as center cell
of the area}
\usage{
h3_flow_dir(inds, z, start_cell)
}
\arguments{
\item{inds}{vector of H3 indexes}

\item{z}{corresponding heights}

\item{start_cell}{center cell of analyzing area}
}
\value{
map of from-to H3 indexes - flow directions
}
\description{
Calculates water flow directions and fills depressions
needs h3 indexes and corresponding heights as well as center cell
of the area
}
