% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{h3_get_base_cell}
\alias{h3_get_base_cell}
\title{Get the base cells of the given H3 indexes.}
\usage{
h3_get_base_cell(h3Str)
}
\arguments{
\item{h3Str}{character vector of H3 indexes}
}
\value{
numeric vector; base cell numbers between 0 and 121
}
\description{
Get the base cells of the given H3 indexes.
}
