% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{h3_get_direct_parents}
\alias{h3_get_direct_parents}
\title{Returns table of input indexes and corresponding parent indexes}
\usage{
h3_get_direct_parents(children_ind)
}
\arguments{
\item{children_ind}{vector of indexes}
}
\value{
table of indexes' pairs
}
\description{
Returns table of input indexes and corresponding parent indexes
}
