% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{h3_get_resolution}
\alias{h3_get_resolution}
\title{Get the resolution of the given H3 indexes.}
\usage{
h3_get_resolution(h3Str)
}
\arguments{
\item{h3Str}{character vector of H3 indexes}
}
\value{
numeric vector
}
\description{
Get the resolution of the given H3 indexes.
}
