% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{h3_global_extremum}
\alias{h3_global_extremum}
\title{Gets stars raster object and returns its min, max or mean value}
\usage{
h3_global_extremum(ind, z, func)
}
\arguments{
\item{func}{}
}
\value{
raster min, max or mean value as double
}
\description{
Gets stars raster object and returns its min, max or mean value
}
