% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{h3_hex_centers_inbbox}
\alias{h3_hex_centers_inbbox}
\title{Defines H3 indexes and center points geographic coordinates
of hexagons which covers defined extent in defined level}
\usage{
h3_hex_centers_inbbox(ext_lon, ext_lat, res)
}
\arguments{
\item{res}{}
}
\value{
map of H3 indexes and corresponding geographic coordinates
}
\description{
Defines H3 indexes and center points geographic coordinates
of hexagons which covers defined extent in defined level
}
