% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{h3_indexes_are_neighbors}
\alias{h3_indexes_are_neighbors}
\title{Check whether the given hexagons are neighbors.}
\usage{
h3_indexes_are_neighbors(origin, destinations)
}
\arguments{
\item{origin}{character scalar; origin H3 index}

\item{destinations;}{character vector of destination H3 indexes}
}
\description{
Check whether the given hexagons are neighbors.
}
