% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{h3_indexes_to_coords}
\alias{h3_indexes_to_coords}
\title{Returns geographic coordinates for vector of hex indexes}
\usage{
h3_indexes_to_coords(inds)
}
\arguments{
\item{inds}{}
}
\value{
map of H3 indexes and corresponding z-values
}
\description{
Returns geographic coordinates for vector of hex indexes
}
