% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{h3_is_res_class_iii}
\alias{h3_is_res_class_iii}
\title{Check whether the given H3 indexes have a resolution
with Class III orientation.}
\usage{
h3_is_res_class_iii(h3Str)
}
\arguments{
\item{h3Str}{character vector of H3 indexes}
}
\value{
logical vector
}
\description{
Check whether the given H3 indexes have a resolution
with Class III orientation.
}
