% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{h3_is_valid}
\alias{h3_is_valid}
\title{Check whether the given H3 indexes are valid.}
\usage{
h3_is_valid(h3Str)
}
\arguments{
\item{h3Str}{character vector of H3 indexes}
}
\value{
logical vector
}
\description{
Check whether the given H3 indexes are valid.
}
