% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_traversal.R
\name{h3_line}
\alias{h3_line}
\title{Get the line of indexes between the given start and end indexes (inclusive).}
\usage{
h3_line(h3_index_start, h3_index_end)
}
\arguments{
\item{h3_index_start}{character scalar; H3 start index}

\item{h3_index_end}{character scalar; H3 end index}
}
\description{
Get the line of indexes between the given start and end indexes (inclusive).
}
