% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{h3_linear_interpolation}
\alias{h3_linear_interpolation}
\title{Gets df with raster's centers coords and corresponding values
and interpolate them into the net of H3 hexagons' centers
which covers the raster with linear Akima interpolator}
\usage{
h3_linear_interpolation(rast_tab, hex_tab)
}
\arguments{
\item{hex_tab}{}
}
\value{
df with hex's centers coords and assigned values (x, y, z)
}
\description{
Gets df with raster's centers coords and corresponding values
and interpolate them into the net of H3 hexagons' centers
which covers the raster with linear Akima interpolator
}
