% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{h3_points_to_H3}
\alias{h3_points_to_H3}
\title{Converts geographic coordinates into H3 indexes.}
\usage{
h3_points_to_H3(lat, lon, res)
}
\arguments{
\item{res}{}
}
\value{
string vector of H3 indices
}
\description{
Converts geographic coordinates into H3 indexes.
}
