% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{h3_raster_to_bbox}
\alias{h3_raster_to_bbox}
\title{Returns raster's extent coordinates in format
which fit to H3 functions}
\usage{
h3_raster_to_bbox(raster_obj)
}
\arguments{
\item{raster_obj}{}
}
\value{
df with columns: ('lon_p', 'lat_p')
}
\description{
Returns raster's extent coordinates in format
which fit to H3 functions
}
