% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{h3_raster_to_hex}
\alias{h3_raster_to_hex}
\title{Gets stars raster object and converts it into H3 indexes}
\usage{
h3_raster_to_hex(rast_obj, h3_level, intrp_method)
}
\arguments{
\item{intrp_method}{}
}
\value{
df with hex's centers coords, assigned raster values (z, x, y)
and H3 indexes
}
\description{
Gets stars raster object and converts it into H3 indexes
}
