% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{h3_resample_down}
\alias{h3_resample_down}
\title{Given parent H3 indexes and corresponding values defines
list of children in any level and maps parents' values}
\usage{
h3_resample_down(level_to, parent_ind, parent_vals)
}
\arguments{
\item{parent_vals}{}
}
\value{
map of H3 indexes and corresponding z-values from parent cells
}
\description{
Given parent H3 indexes and corresponding values defines
list of children in any level and maps parents' values
}
