% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{h3_resample_up}
\alias{h3_resample_up}
\title{Returns table of input indexes and corresponding parent indexes}
\usage{
h3_resample_up(func, children_ind, children_vals)
}
\arguments{
\item{func}{aggregation function (sum, max, avg, majority)}

\item{children_ind}{vector of indexes}

\item{children_vals}{vector of values in the indexes}
}
\value{
map of parents' H3 indexes and corresponding z-values
}
\description{
Returns table of input indexes and corresponding parent indexes
}
