% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{h3_simple_focal}
\alias{h3_simple_focal}
\title{Returns hexs after simple focal operations, without matrices}
\usage{
h3_simple_focal(inds, z, stat_type, vecinity)
}
\arguments{
\item{vecinity}{}
}
\value{
map of H3 indexes and corresponding z-values
}
\description{
Returns hexs after simple focal operations, without matrices
}
