% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{h3_simple_sum}
\alias{h3_simple_sum}
\title{Calculates sum of values in equal H3 cells of two rasters
- vectors with indexes and values; resamples one raster if
they have different levels}
\usage{
h3_simple_sum(ind1, z1, ind2, z2)
}
\arguments{
\item{z2}{}
}
\value{
map of H3 indexes and corresponding z-values sum
}
\description{
Calculates sum of values in equal H3 cells of two rasters
- vectors with indexes and values; resamples one raster if
they have different levels
}
