% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_hierachy.R
\name{h3_to_children}
\alias{h3_to_children}
\title{Get the children of the given hexagon at a particular resolution.}
\usage{
h3_to_children(h3_index, res)
}
\arguments{
\item{h3_index}{character scalar representing a valid H3 index}

\item{res}{resolution of children}
}
\value{
character vector
}
\description{
Get the children of the given hexagon at a particular resolution.
}
