% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_hierachy.R
\name{h3_to_parent}
\alias{h3_to_parent}
\title{Get the parent of the given hexagon at a particular resolution.}
\usage{
h3_to_parent(h3_index, res)
}
\arguments{
\item{h3_index}{character scalar representing a valid H3 index}

\item{res}{resolution of parent}
}
\value{
character scalar
}
\description{
Get the parent of the given hexagon at a particular resolution.
}
