% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{h3_unidirectional_edge_is_valid}
\alias{h3_unidirectional_edge_is_valid}
\title{Check whether the given indexes are valid H3 edge indexes.}
\usage{
h3_unidirectional_edge_is_valid(h3_edge_indexes)
}
\arguments{
\item{h3_edge_indexes}{character vector of H3 edge indexes}
}
\description{
Check whether the given indexes are valid H3 edge indexes.
}
