% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{h3_zonal_sum}
\alias{h3_zonal_sum}
\title{Calculates sum of values in H3 cells of the second raster which
corresponds to H3 cells in the first raster's zones, i.e. cells
that share the same value; resamples one raster if
they have different levels}
\usage{
h3_zonal_sum(zone_ind, zone_z, rast_ind, rast_z)
}
\arguments{
\item{rast_z}{}
}
\value{
map of H3 indexes, corresponding zone codes from first raster
and sums in each zone
}
\description{
Calculates sum of values in H3 cells of the second raster which
corresponds to H3 cells in the first raster's zones, i.e. cells
that share the same value; resamples one raster if
they have different levels
}
