% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hex_area}
\alias{hex_area}
\title{Average hexagon area in square meters or kilometers at the given resolution.}
\usage{
hex_area(res, unit)
}
\arguments{
\item{res}{resolution between 0 and 15}

\item{unit}{either \code{m2} or \code{km2}}
}
\value{
numeric vector
}
\description{
Average hexagon area in square meters or kilometers at the given resolution.
}
