% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_traversal.R
\name{hex_ring}
\alias{hex_ring}
\title{Get the hollow hexagonal ring centered at the given origin.}
\usage{
hex_ring(h3_index, radius = 1)
}
\arguments{
\item{h3_index}{character vector of H3 indexes}

\item{radius}{number of rings around the given center}
}
\value{
character vector of H3 indexes
}
\description{
Get the hollow hexagonal ring centered at the given origin.
}
