% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_traversal.R
\name{k_ring}
\alias{k_ring}
\title{Get all hexagons in a k-ring around the given center.}
\usage{
k_ring(h3_index, radius = 1)
}
\arguments{
\item{h3_index}{character vector of H3 indexes}

\item{radius}{number of rings around the given center}
}
\value{
character vector of H3 indexes
}
\description{
Get all hexagons in a k-ring around the given center.
}
\note{
The order of the hexagons is undefined.
}
