% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_traversal.R
\name{k_ring_distances}
\alias{k_ring_distances}
\title{Get all hexagons in a k-ring around a given center ordered by distance from the origin.}
\usage{
k_ring_distances(h3_index, radius = 1)
}
\arguments{
\item{h3_index}{character vector of H3 indexes}

\item{radius}{number of rings around the given center}
}
\value{
tibble with 2 variables: \code{h3_index} and \code{distance}
}
\description{
Get all hexagons in a k-ring around a given center ordered by distance from the origin.
}
\note{
The order of the hexagons within each ring is undefined.
}
