% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{num_hexagons}
\alias{num_hexagons}
\title{Number of unique H3 indexes at the given resolution.}
\usage{
num_hexagons(res)
}
\arguments{
\item{res}{numeric vector; resolution between 0 and 15}
}
\value{
numeric vector
}
\description{
Number of unique H3 indexes at the given resolution.
}
