% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfill.R
\name{polyfill}
\alias{polyfill}
\alias{polyfill.matrix}
\alias{polyfill.sfc}
\alias{polyfill.sf}
\title{Get all hexagons with centers contained in a given polygon.}
\usage{
polyfill(polygon, res = 7)

\method{polyfill}{matrix}(polygon, res = 7)

\method{polyfill}{sfc}(polygon, res = 7)

\method{polyfill}{sf}(polygon, res = 7)
}
\arguments{
\item{polygon}{geo-coordinates of polygon as lat/lng pairs (matrix) or object of class \code{sf}
or \code{sfc}}

\item{res}{resolution between 0 and 15}
}
\value{
character vector of H3 indexes
}
\description{
Get all hexagons with centers contained in a given polygon.
}
\examples{
nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"),  quiet = TRUE) \%>\%
  sf::st_transform(4326)
h3_indexes <- polyfill(nc[1, ], res = 7)
}
