% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_indexing.R
\name{h3_to_geo}
\alias{h3_to_geo}
\title{Get the centers of the given H3 indexes.}
\usage{
h3_to_geo(h3_index)
}
\arguments{
\item{h3_index}{character vector of H3 indexes}
}
\description{
Get the centers of the given H3 indexes.
}
