% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_indexing.R
\name{h3_to_geo_boundary_sf}
\alias{h3_to_geo_boundary_sf}
\title{Get the polygons of the given H3 indexes}
\usage{
h3_to_geo_boundary_sf(h3index)
}
\arguments{
\item{h3index}{character vector of H3 indexes}
}
\value{
object of class \code{sf}
}
\description{
Get the polygons of the given H3 indexes
}
