% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfill.R
\name{polyfill}
\alias{polyfill}
\alias{polyfill.matrix}
\alias{polyfill.sfc}
\alias{polyfill.sf}
\title{Get all hexagons with centers contained in a given polygon.}
\usage{
polyfill(polygon, res = 7)

\method{polyfill}{matrix}(polygon, res = 7)

\method{polyfill}{sfc}(polygon, res = 7)

\method{polyfill}{sf}(polygon, res = 7)
}
\arguments{
\item{polygon}{geo-coordinates of polygon as lng/lat pairs (matrix) or object of class \code{sf}
or \code{sfc}}

\item{res}{resolution between 0 and 15}
}
\description{
Get all hexagons with centers contained in a given polygon.
}
